do not split files without need
