--- bpf/include/lib/config.h
@@ -1,39 +1,0 @@
-/*
-Copyright 2025 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-#pragma once
-
-// all constants here are overriden from user-space level
-volatile const uint32_t IPV4_NAT_PREFIX = 0;
-volatile const uint32_t IPV4_NAT_MASK = 0;
-
-volatile const uint32_t IPV6_NAT_PREFIX_0 = 0;
-volatile const uint32_t IPV6_NAT_PREFIX_1 = 0;
-volatile const uint32_t IPV6_NAT_PREFIX_2 = 0;
-
-volatile const uint32_t IPV6_NAT_MASK_0 = 0;
-volatile const uint32_t IPV6_NAT_MASK_1 = 0;
-volatile const uint32_t IPV6_NAT_MASK_2 = 0;
-
-volatile const uint32_t POD_PREFIX_0 = 0;
-volatile const uint32_t POD_PREFIX_1 = 0;
-volatile const uint32_t POD_PREFIX_2 = 0;
-volatile const uint32_t POD_PREFIX_3 = 0;
-
-volatile const uint32_t POD_MASK_0 = 0;
-volatile const uint32_t POD_MASK_1 = 0;
-volatile const uint32_t POD_MASK_2 = 0;
-volatile const uint32_t POD_MASK_3 = 0;
