--- bpf/include/lib/csum.h
@@ -1,30 +1,0 @@
-/*
-Copyright 2025 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-#pragma once
-
-#include <linux/types.h>
-#include <bpf/bpf_endian.h>
-
-static __always_inline __wsum csum_add(__wsum csum, __wsum addend) {
-	csum += addend;
-	return csum + (csum < addend);
-}
-
-static __always_inline __wsum csum_sub(__wsum csum, __wsum addend) {
-	return csum_add(csum, ~addend);
-}
-
